# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VizTR Super Admin API",
    "version": "1.0.0",
    "description": "Super Admin Dashboard API for VizTR Architecture Visualization Platform",
    "contact": {
      "name": "VizTR Team",
      "email": "api@viztr.studio"
    }
  },
  "servers": [
    {
      "url": "https://api.viztr.studio",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/contact": {
      "get": {
        "summary": "List contact submissions",
        "parameters": [
          {
            "schema": {
              "type": "object",
              "properties": {
                "search": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "new",
                    "read",
                    "responded",
                    "archived"
                  ]
                },
                "service_interest": {
                  "type": "string",
                  "enum": [
                    "Architectural",
                    "Virtual Reality",
                    "Pixel Streaming",
                    "WebXR",
                    "WebAR",
                    "Virtual Tour 360",
                    "Animation",
                    "Still Renders",
                    "Other"
                  ]
                },
                "date_from": {
                  "type": "string",
                  "format": "date"
                },
                "date_to": {
                  "type": "string",
                  "format": "date"
                },
                "page": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "default": 1
                },
                "pageSize": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 100,
                  "default": 20
                },
                "sortBy": {
                  "type": "string"
                },
                "sortOrder": {
                  "type": "string",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              }
            },
            "required": false,
            "in": "query",
            "name": "ContactFilters"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated contact submissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "phone": {
                            "type": "string"
                          },
                          "company": {
                            "type": "string"
                          },
                          "service_interest": {
                            "type": "string",
                            "enum": [
                              "Architectural",
                              "Virtual Reality",
                              "Pixel Streaming",
                              "WebXR",
                              "WebAR",
                              "Virtual Tour 360",
                              "Animation",
                              "Still Renders",
                              "Other"
                            ]
                          },
                          "message": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "new",
                              "read",
                              "responded",
                              "archived"
                            ]
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "email",
                          "status",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "by_status": {
                          "type": "object",
                          "properties": {
                            "new": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "read": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "responded": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "archived": {
                              "type": "integer",
                              "minimum": 0
                            }
                          }
                        },
                        "by_service_interest": {
                          "type": "object",
                          "properties": {
                            "Architectural": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Virtual Reality": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Pixel Streaming": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "WebXR": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "WebAR": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Virtual Tour 360": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Animation": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Still Renders": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "Other": {
                              "type": "integer",
                              "minimum": 0
                            }
                          }
                        }
                      },
                      "required": [
                        "total",
                        "by_status",
                        "by_service_interest"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "total",
                    "page",
                    "pageSize",
                    "totalPages",
                    "stats"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit contact form",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "company": {
                    "type": "string"
                  },
                  "service_interest": {
                    "type": "string",
                    "enum": [
                      "Architectural",
                      "Virtual Reality",
                      "Pixel Streaming",
                      "WebXR",
                      "WebAR",
                      "Virtual Tour 360",
                      "Animation",
                      "Still Renders",
                      "Other"
                    ]
                  },
                  "message": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact submitted"
          }
        }
      }
    },
...
  },
...
}